ceate route between card and cardDetails
