+++ src/pages/Home.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Header from '../components/navbar/Header'
-import Products from '../components/navbar/products';
+import Products from '../components/products';
 
 const Home = () => (
   <div className="container">
